% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Create a connection to the database}
\usage{
connect_to_database()
}
\value{
Returns an S4 object. This object is used to communicate with the database engine.
}
\description{
Creates a connection to the database that can be called later if this function
call is stored in an object.
}
