% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_to_production_database}
\alias{connect_to_production_database}
\title{Create a connection to Porowhita Hauwhā}
\usage{
connect_to_production_database(read_only = TRUE)
}
\arguments{
\item{read_only}{Is the connection to the database read-only (\code{TRUE}) or
not (\code{FALSE})? Default is \code{TRUE}.}
}
\value{
An S4 object. This object is used to communicate with the
database engine.
}
\description{
Creates a read-only connection to the live version of the Porowhita Hauwhā
database.
}
