% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_to_test_database}
\alias{connect_to_test_database}
\title{Create a connection to a test version of Porowhita Hauwhā}
\usage{
connect_to_test_database()
}
\value{
An S4 object. This object is used to communicate with the
database engine.
}
\description{
Create a connection to a test version of Porowhita Hauwhā
}
