% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_to_writable_database}
\alias{connect_to_writable_database}
\title{Create a connection to Porowhita Hauwhā}
\usage{
connect_to_writable_database(test_db = FALSE)
}
\arguments{
\item{test_db}{Is this a connection to the test database? Defaults
to \code{FALSE}.}
}
\value{
An S4 object. This object is used to communicate with the
database engine, either live or test.
}
\description{
Creates a connection to either the live version or the test version
of the Porowhita Hauwhā database. This can be used later if this function call
is stored in an object. The connection to the database is not read-only,
and as a result only one such connection is allowed at a time.
}
