% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{create_test_database}
\alias{create_test_database}
\title{Create a database with test data}
\usage{
create_test_database(test_conn)
}
\arguments{
\item{test_conn}{The database instance used to connect to the in-memory database.}
}
\value{
An S4 object. This object is used to communicate with the
test database engine.
}
\description{
Replicates the structure of Porowhita HauwhÄ but fills it with a small sample
of test data.
}
