% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{disconnect_from_database}
\alias{disconnect_from_database}
\title{Disconnect from the database}
\usage{
disconnect_from_database(conn, test_db = FALSE, confirm = TRUE)
}
\arguments{
\item{conn}{The database instance that was called using \code{connect_to_database()}}

\item{test_db}{Are you attempting to disconnect from the test database? Defaults to FALSE.}

\item{confirm}{Should the function print a confirmation message to the console? Defaults to TRUE.}
}
\value{
An optional confirmation statement printed to the console.
}
\description{
Explicitly shuts down the database instance associated with the connection, \code{conn}.
}
\examples{
con <- connect_to_database(test_db = TRUE)
disconnect_from_database(con, test_db = TRUE)
}
