% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_assets}
\alias{get_assets}
\title{Retrieve a list of all assets}
\usage{
get_assets(..., con = connect_to_database())
}
\arguments{
\item{...}{Optional expressions to filter the asset list, defied in terms of the
variables in the returned tibble.}

\item{con}{The database instance. Defaults to what is returned from calling
\code{connect_to_database()}.}
}
\value{
A tibble with six variables.
}
\description{
Retrieve a list of assets from PorowhitÄ Hauwha. An asset is a structure that
provides shelter for the public.
}
\examples{
# Retrieve a list of assets that are rural halls in the Rodney Local Board (code not run)
# get_assets(designation == "Rural Hall" & local_board == "Rodney")
}
