% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_community_spaces}
\alias{get_community_spaces}
\title{Return a dataframe of community spaces}
\usage{
get_community_spaces(test_db = FALSE)
}
\arguments{
\item{test_db}{Retrieve this data from the test database? Defaults to
\code{FALSE}.}
}
\value{
A tibble with one community space per row, and one column per
attribute for each community space.
}
\description{
Return a dataframe of community spaces
}
