% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_entities}
\alias{get_entities}
\title{Retrieve a list of all entities}
\usage{
get_entities(..., test_db = FALSE)
}
\arguments{
\item{...}{Optional expressions to filter the entities list, defied in terms of
the variables in the returned tibble.}

\item{test_db}{Retrieve this data from the test database? Defaults to FALSE.}
}
\value{
A tibble with eight variables.
}
\description{
Retrieve a list of entities from Porowhita HauwhÄ. An entity is a construct,
assigned by Auckland Council, that groups together assets, spaces, or
services.
}
\examples{
# Retrieve a list of entities that are classed as Community Hubs from the test database
get_entities(designation == "Community Hub", test_db = TRUE)
}
