% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_facilities}
\alias{get_facilities}
\title{Retrieve a list of all facilities}
\usage{
get_facilities(..., test_db = FALSE)
}
\arguments{
\item{...}{List of attribute-value pairs with which to filter the list of
returned facilities.}

\item{test_db}{Retrieve this data from the test database? Defaults to
\code{FALSE}.}
}
\value{
A tibble with one facility per row, and one column per attribute for
each facility.
}
\description{
Returns a list of all facilities in the database, with key information about
each. For more information about how facilities are classified in the
database, please consult the documentation that accompanies the test data.
}
\examples{
# Return a list of all community-led facilities in the database
get_facilities(delivery_model == "Community-led facility", test_db = TRUE)
}
