% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_facilities_attributes_audit}
\alias{get_facilities_attributes_audit}
\title{Retrieve a list of facilities audit entries}
\usage{
get_facilities_attributes_audit(..., test_db = FALSE)
}
\arguments{
\item{...}{Optional expressions to filter the audit list, defined in terms of
the variables in the returned tibble.}

\item{test_db}{Retrieve this data from the test database? Defaults to FALSE.}
}
\value{
A tibble with eight columns.
}
\description{
Retrieve a list of audit entries from Porowhita Hauwhā. The facilities_attributes_audit_table
is used for tracking changes (inserts, updates, deletes) to the attributes of facilities over time.
}
\examples{
# Retrieve a list of audit entries that describe a changed attribute for a facility in
the facilities_attributes table.
get_facilities_attributes_audit()
}
