% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_facilities_names}
\alias{get_facilities_names}
\title{Retrieve a list of all facility names}
\usage{
get_facilities_names()
}
\value{
A tibble with four columns.
}
\description{
Retrieve a list of facility names, variants of the names, and an ID and
facility type classification (asset, space, or entity) for retrieving attributes.
}
