% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_file_path}
\alias{get_file_path}
\title{Retrieve the full path to a file in File Storage.}
\usage{
get_file_path(file_name)
}
\arguments{
\item{file_name}{The name of the file for which you're retrieving the full path.}
}
\value{
A character string containing a file path.
}
\description{
Reads your .Renviron file to get the path to the team's SharePoint Document Library,
as sync'd to your local machine. Appends to this the name of the file,
thus returning the full path to the file to be used elsewhere.
}
