% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_libraries}
\alias{get_libraries}
\title{Return a dataframe of community libraries}
\usage{
get_libraries(test_db = FALSE)
}
\arguments{
\item{test_db}{Retrieve this data from the test database? Defaults to
\code{FALSE}.}
}
\value{
A tibble with one library per row, and one column per attribute for
each library.
}
\description{
Return a dataframe of community libraries
}
