% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_names}
\alias{get_names}
\title{Find a facility name in the database}
\usage{
get_names(names, test_db = FALSE)
}
\arguments{
\item{names}{The facility name(s) that you're looking for in the database}

\item{test_db}{Retrieve this data from the test database (\code{TRUE}) or not(\code{FALSE}). Defaults to \code{FALSE}.}
}
\value{
A tibble.
}
\description{
Find a facility name in the database
}
