% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_new_entry}
\alias{get_new_entry}
\title{Return the record that was just added to a database table}
\usage{
get_new_entry(conn, tbl_name, new_id)
}
\arguments{
\item{conn}{The database instance used to connect to the database.}

\item{tbl_name}{The name of the table to which a record was just added.}

\item{new_id}{The ID of the record that was just added to the table.}
}
\value{
A data frame with one row.
}
\description{
Return the record that was just added to a database table
}
