% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_new_id}
\alias{get_new_id}
\title{Create a new ID for a table in the database}
\usage{
get_new_id(conn, tbl_name)
}
\arguments{
\item{conn}{The database instance used to connect to the database.}

\item{tbl_name}{The name of the database table.}
}
\value{
A character string
}
\description{
Removes any letters from the values in the field called \code{id}, converts the
values to numeric, then extracts the maximum value and returns it as a character.
}
