% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_spaces}
\alias{get_spaces}
\title{Retrieve a list of all spaces}
\usage{
get_spaces(..., test_db = FALSE)
}
\arguments{
\item{...}{Optional expressions to filter the space list, defied in terms of
the variables in the returned tibble.}

\item{test_db}{Retrieve this data from the test database? Defaults to FALSE.}
}
\value{
A tibble with eight variables.
}
\description{
Retrieve a list of spaces from Porowhita Hauwhā. A space is an area in which
services are delivered to the public, located within an asset
}
\examples{
# Retrieve a list of spaces that are classed as community libraries from the test database
get_spaces(designation == "Community Library", test_db = TRUE)
}
