% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_change_log}
\alias{insert_change_log}
\title{Add An Item to the Facilities Change Log}
\usage{
insert_change_log(
  facility_attribute_id,
  facility_type,
  facility_id,
  attribute,
  value,
  valid_from = NA,
  valid_to = NA,
  notes = NULL,
  test_db = FALSE
)
}
\arguments{
\item{facility_attribute_id}{The value of the facility_attribute_id for the
facility whose change you're recording.}

\item{facility_type}{Is the facility an asset, a space, or an entity?}

\item{facility_id}{The value of the facility_id field for the facility whose
change you're recording}

\item{attribute}{The attribute, or field name, that is affected by the
change.}

\item{value}{The new value that is being supplied for this facility.}

\item{valid_from}{From when did this change take effect?}

\item{valid_to}{Until when is this change valid?}

\item{notes}{Explanatory notes for the change, such as the justification for
the change, the meaning of the change, and/or the person who authorised
this change.}

\item{test_db}{Is this change being applied to a facility in the test
database (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{FALSE}.}
}
\value{
A tibble with 1 row and 9 columns showing the log entry you supplied
as recorded in the database.
}
\description{
Inserts a new row into the facilities_attributes_bridge_table, where changes
to the various facilities tables are recorded.
}
\examples{
change <- insert_change_log(
facility_attribute_id = "FA601",
facility_type = "Space",
facility_id = "S374",
attribute = "bookable",
value = "Y",
valid_from = Sys.Date(),
notes = "This room can now be booked by members of the public.",
test_db = TRUE
)

change
}
