% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_contact}
\alias{insert_contact}
\title{Add a new partner contact into the database}
\usage{
insert_contact(
  name = NA,
  partner_id = NA,
  role = c("Site contact", "Chairperson", "Reporting person"),
  email_address = NA,
  phone_number = NA,
  test_db = FALSE
)
}
\arguments{
\item{name}{The name of the partner contact that is being added to the
database. The first name and surname of the contact should be provided.}

\item{partner_id}{The record ID of the partner for whom this individual a
contact. The joins the contact to the \code{partners} table appropriately.}

\item{role}{The role of this contact, either \verb{Site contact}, \code{Chairperson},
or \verb{Reporting provider}. Note that multiple roles might be assigned to one
individual, with a separate record in \code{contacts} for each of these roles.}

\item{email_address}{An optional email address for the contact. Ideally,
either an email address or a telephone number should be provided (or both).}

\item{phone_number}{An optional phone number for the contact. Ideally, either
an email address or a telephone number should be provided (or both).}

\item{test_db}{Is this connection to the test database (\code{TRUE}) or not
(\code{FALSE})? Defaults to \code{FALSE}.}
}
\value{
A data frame with one row containing the newly-added entry.
}
\description{
If supplied with at least a partner_id and name of the contact, create a new
record for a partner contact in the database.
}
\examples{
insert_contact(name = "Bilbo Baggins", partner_id = "P1", role = "Chairperson", test_db = TRUE)
}
