% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_name}
\alias{insert_name}
\title{Add a new facility name to the database}
\usage{
insert_name(
  new_name = NULL,
  role = c("alternate", "primary"),
  facilities_attributes_id = NULL,
  test_db = FALSE
)
}
\arguments{
\item{new_name}{The name of the facility that is being added.}

\item{role}{Is this an alternate name for the facility, or its primary name? Defaults to alternate.}

\item{facilities_attributes_id}{The matching ID in the \code{facilities_attributes} table for the facility whose name is being supplied.}

\item{test_db}{Is this connection to the test database (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{FALSE}.}
}
\value{
A tibble with one row containing the newly-added record.
}
\description{
Given a name and an ID for that facility in the \code{facilities_attributes}
table, insert a new record into the \code{names} table.
}
\examples{
insert_name(
    new_name = "Hobbiton Hall",
    facilities_attributes_id = "FA174",
    test_db = TRUE
)
}
