% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_partner}
\alias{insert_partner}
\title{Add a new partner into the database}
\usage{
insert_partner(
  name = NULL,
  type = c("Charitable Trust", "Incorporated Society", "Other"),
  facility_owner = TRUE,
  service_provider = TRUE,
  legal_status_number = NULL,
  test_db = FALSE
)
}
\arguments{
\item{name}{The name of the partner that is being added to the database. This is the minimum information that should be supplied.}

\item{type}{Is this partner a Charitable Trust, an Incorporated Society, or Other? Default to Charitable Trust.}

\item{facility_owner}{Does this partner own the facility in which they operate (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{TRUE}.}

\item{service_provider}{Does this partner provide services (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{TRUE}.}

\item{legal_status_number}{The Legal Status Number for this partner, if available.}

\item{test_db}{Is this connection to the test database (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{FALSE}.}
}
\value{
A data frame with one row containing the newly-added entry.
}
\description{
If supplied with at least a name, create a new record for a partner in
the database.
}
\examples{
insert_partner(
  name = "Hobbiton Community Trust",
  facility_owner = FALSE,
  test_db = TRUE,
(

}
