% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_record}
\alias{insert_record}
\title{Add a new record into a database table}
\usage{
insert_record(..., test_db = FALSE, new_id_prefix = NULL, tbl_name = NULL)
}
\arguments{
\item{...}{List of values to add include in the new record.}

\item{test_db}{Is this connection to the test database (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{FALSE}.}

\item{new_id_prefix}{An optional ID prefix for the new record (e.g. "A" in the assets table).}

\item{tbl_name}{The name of the table into which this record is being inserted.}
}
\value{
a tibble with one row containing the newly-inserted record.
}
\description{
This is a generic function for inserting a new record into one of the tables
in the database.
}
