% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porowhitahauwha-package.R
\docType{package}
\name{porowhitahauwha-package}
\alias{porowhitahauwha}
\alias{porowhitahauwha-package}
\title{porowhitahauwha: Interact With Data About Facilities and Partners Within the Connected Communities Portfolio}
\description{
This package allows authorised staff at Auckland Council to read from and write to Porowhita Hauwhā, a database that stores information about facilities, partners, and related data that falls within the Connected Communities portfolio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbinAC/porowhitahauwha}
  \item Report bugs at \url{https://github.com/lddurbinAC/porowhitahauwha/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{lee.durbin@aucklandcouncil.govt.nz}

Other contributors:
\itemize{
  \item Auckland Council [copyright holder]
}

}
\keyword{internal}
