% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_assets}
\alias{test_assets}
\title{Porowhita Hauwh훮 test data: assets}
\format{
\subsection{'test_assets'}{

A data frame with 2 rows and 9 columns:
\describe{
\item{id}{Unique identifier. This is the primary key for the table.}
\item{name}{The primary name of the asset.}
\item{local_board}{The Local Board in which this asset is located.}
\item{asset_type}{Is this asset a standalone building, a vehicle, or some other type of structure?}
\item{physical_address}{The physical address of the asset, including street name and number (where available).}
\item{latitude}{What is the latitude of this asset?}
\item{longitude}{What is the longitude of this asset?}
\item{image}{A url to an image of this asset, if available.}
}
}
}
\usage{
test_assets
}
\description{
A subset of the data stored in the Porowhit훮 Hauwha table called \code{assets}.

An asset is a physical structure that provides shelter for the public, typically
a building.

This data has been made available for use in package documentation examples,
and for the user to explore the kind of information stored in the live
Porowhit훮 Hauwh훮 database.
}
\keyword{datasets}
