% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_contacts}
\alias{test_contacts}
\title{Porowhita Hauwhā test data: test_contacts}
\format{
\subsection{'test_contacts'}{

A data frame with 1 row and 6 columns:
\describe{
\item{id}{Unique identifier. This is the primary
key for the table.}
\item{partner_id}{Foreign key for joining to the \code{partners} table.}
\item{name}{The name of the contacts for partners}
\item{email_address}{The email_address of the contacts for partners}
\item{phone_number}{The phone_number of the contacts for partners}
\item{role}{The role of the contacts for partners}
}
}
}
\usage{
test_contacts
}
\description{
A subset of the data stored in the Porowhitā Hauwha table called
\code{contacts}.
}
\keyword{datasets}
