% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_entities}
\alias{test_entities}
\title{Porowhita Hauwh훮 test data: entities}
\format{
\subsection{'test_entities'}{

A data frame with 1 row and 2 columns:
\describe{
\item{id}{Unique identifier. This is the primary key for the table.}
\item{name}{The primary name of the entity.}
}
}
}
\usage{
test_entities
}
\description{
A subset of the data stored in the Porowhit훮 Hauwha table called \code{entities}.

An entity is a construct, assigned by Auckland Council, that groups together
assets, spaces, or services.

This data has been made available for use in package documentation examples,
and for the user to explore the kind of information stored in the live
Porowhit훮 Hauwh훮 database.
}
\keyword{datasets}
