% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_entity_bridge_table}
\alias{test_entity_bridge_table}
\title{Porowhita Hauwh훮 test data: entity_bridge_table}
\format{
\subsection{'test_entity_bridge_table'}{

A data frame with 4 rows and 7 columns:
\describe{
\item{id}{Unique identifier. This is the primary key for the table.}
\item{facility_type}{Describes whether this facility is an asset, a space, or an entity.}
\item{facility_id}{A foreign key for joining to the relevant facilities table, as defined by \code{facility_type}.}
\item{entity_id}{A foreign key for joining to the \code{entities} table.}
\item{valid_from}{A date defining a point in time from which this relationship was established.}
\item{valid_to}{A date defining a point in time from which this relationship was disestablished.}
\item{notes}{An explanation for why this relationship was disestablished.}
}
}
}
\usage{
test_entity_bridge_table
}
\description{
A subset of the data stored in the Porowhit훮 Hauwha table called \code{entity_bridge_table}.

An entity is composed of at least two records in either the assets, spaces, or services
tables. This table joins the entity to the assets, spaces, or service from which
it is formed.

This data has been made available for use in package documentation examples,
and for the user to explore the kind of information stored in the live
Porowhit훮 Hauwh훮 database.
}
\keyword{datasets}
