% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_facilities_attributes}
\alias{test_facilities_attributes}
\title{Porowhita Hauwhā test data: facilities_attributes}
\format{
\subsection{'test_facilities_attributes'}{

A data frame with 5 rows and 9 columns:
\describe{
\item{id}{Unique identifier. This is the primary key for the table.}
\item{property_code}{}
\item{provider_legal_status_number}{}
\item{entry_access_type}{}
\item{facility_type}{Describes whether this facility is an asset, a space, or an entity.}
\item{facility_id}{A foreign key for joining to the relevant facilities table, as defined by \code{facility_type}.}
\item{designation}{Describes the function of this facility, for example "Rural Hall", "Community Library", "Room" etc.}
\item{delivery_model}{Categorises the facility as either community-led or Council-led. A community-led facility is operated by community partners.}
\item{facility_ownership}{Describes whether the facility is owned by Council, by Kainga Ora, or is privately owned.}
\item{staffed}{Is the facility staffed during opening hours (\code{TRUE}) or not (\code{FALSE})?}
\item{closed}{Is the facility currently closed to the public (\code{TRUE}) or not (\code{FALSE})?}
\item{leased}{Is there a community lease agreement in place for this facility (\code{TRUE}) or not (\code{FALSE})?}
}
}
}
\usage{
test_facilities_attributes
}
\description{
A subset of the data stored in the Porowhitā Hauwha table called \code{facilities_attributes}.

The need for a separate table to store information (or attributes) about
facilities is because assets, spaces, and entities have common fields; managing
those fields here simplifies the process of managing the data.

This data has been made available for use in package documentation examples,
and for the user to explore the kind of information stored in the live
Porowhita Hauwhā database.
}
\keyword{datasets}
