% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_names}
\alias{test_names}
\title{Porowhita Hauwh훮 test data: names}
\format{
\subsection{'test_names'}{

A data frame with 7 rows and 4 columns:
\describe{
\item{id}{Unique identifier. This is the primary key for the table.}
\item{value}{The name of the facility, categorised by the \code{role} field.}
\item{role}{Categorises the name stored in \code{value} as either the primary or alternate name used for this facility. A facility can have multiple alternate names.}
\item{facilities_attributes_id}{A foreign key for joining to the \code{facilities_attributes} table, where more information about this facility is found.}
}
}
}
\usage{
test_names
}
\description{
A subset of the data stored in the Porowhit훮 Hauwha table called \code{names}.

Each facility, whether an asset, a space, or an entity, should have at least
one record in this table: a name of role "primary". Some facilities also have
at least one name of role "alternate" where we're aware of any alternate names.
The primary name is also the name used in the relevant table, whether an asset,
a space, or an entity.

This data has been made available for use in package documentation examples,
and for the user to explore the kind of information stored in the live
Porowhit훮 Hauwh훮 database.
}
\keyword{datasets}
