% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_partners_bridge_table}
\alias{test_partners_bridge_table}
\title{Porowhita Hauwhā test data: partners_bridge_table}
\format{
\subsection{'test_partners_bridge_table'}{

A data frame with 1 row and 4 columns:
\describe{
\item{id}{Unique identifier. This is the primary
key for the table.}
\item{partner_id}{Foreign key for joining to the \code{partners} table.}
\item{facility_type}{Is the facility an
asset, a space, or an entity?}
\item{facility_id}{Foreign key for joining
on the assets, spaces, or entities table.}
}
}
}
\usage{
test_partners_bridge_table
}
\description{
A subset of the data stored in the Porowhitā Hauwha table called
\code{partners_bridge_table}.

This table records the relationship between partners and facilities. Not
every partner is associated with a facility, but the majority who are will
be included here. Full details on partners are found in the \code{partners}
table, and full details on facilities can be found either in \code{assets},
\code{spaces}, or \code{entities}, together with their common attributes
recorded in the \code{facilities_attributes} table.
}
\keyword{datasets}
