% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_spaces}
\alias{test_spaces}
\title{Porowhita Hauwhā test data: spaces}
\format{
\subsection{'test_spaces'}{

A data frame with 4 rows and 6 columns:
\describe{
\item{id}{Unique identifier. This is the primary key for the table.}
\item{name}{The primary name of the space.}
\item{asset_id}{A foreign key for joining spaces with assets.}
\item{bookable}{Is this a bookable space? TRUE if it is, FALSE if not.}
\item{booking_method}{If this is a bookable space, how is it booked? For example, "Sphere".}
\item{image}{A url to an image of this asset, if available.}
}
}
}
\usage{
test_spaces
}
\description{
A subset of the data stored in the Porowhitā Hauwha table called \code{spaces}.
A space is a an area in which services are delivered to the public; a space
is typically a room.

This data has been made available for use in package documentation examples,
and for the user to explore the kind of information stored in the live
Porowhita Hauwhā database.
}
\keyword{datasets}
