% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_facility}
\alias{update_facility}
\title{Update a Facility in the Database}
\usage{
update_facility(facility_name = NULL, field, value, test_db = FALSE)
}
\arguments{
\item{facility_name}{The name of the facility. Must be the primary name.}

\item{field}{A list of field names whose values you wish to change. Should
align with \code{value}.}

\item{value}{A list of values you wish to change. Should align with
\code{field}.}

\item{test_db}{Is this a connection to the test database? Defaults to
\code{FALSE}.}
}
\value{
A tibble with one row containing the record of the updated facility.
}
\description{
Given a facility name, a list of fields, and a corresponding list of values,
update facility information in the database.
}
\examples{
  update_facility(
  facility_name = "Buckland Community Centre",
  field = list("physical_address", "designation"),
  value = list("Corner Logan and Buckville Road", "Community Centre"),
  test_db = TRUE
  )
}
