% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Create a connection to the database}
\usage{
connect_to_database(db_file = "porowhita_hauwha.duckdb", read_only = TRUE)
}
\arguments{
\item{db_file}{The name of the .duckdb file. By default assumes this is called
"Porowhita Hauwhā".}

\item{read_only}{Declares if the connection to the database is read-only.
If \code{read_only = TRUE} (the default), the connection is read-only and multiple
R processes can access the same database file at the same time.}
}
\value{
An S4 object. This object is used to communicate with the
database engine.
}
\description{
Creates a read-only connection to a DuckDB database that can be used later if
this function call is stored in an object.
}
\examples{
connect_to_database(db_file = ":memory:")
}
