% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Create a connection to Porowhita Hauwhā.}
\usage{
connect_to_database(test_db = FALSE)
}
\arguments{
\item{test_db}{Is this a connection to the test database? Defaults to FALSE.}
}
\value{
An S4 object. This object is used to communicate with the
database engine, either live or test.
}
\description{
Creates a connection to either the live version or the test version
of the Porowhita Hauwhā database. This can be used later if this function call
is stored in an object. The connection to the live database is read-only.
}
\examples{
test_conn <- connect_to_database(test_db = TRUE)
dplyr::tbl(test_conn, "assets")
}
