% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{disconnect_from_database}
\alias{disconnect_from_database}
\title{Disconnect from the database}
\usage{
disconnect_from_database(con)
}
\arguments{
\item{con}{The database instance that was called using \code{connect_to_database}}
}
\value{
A confirmation statement printed to the console.
}
\description{
Explicitly shuts down the database instance associated with the connection, \code{con}.
}
\examples{
con <- connect_to_database(":memory:")
disconnect_from_database(con)
}
