% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_assets}
\alias{get_assets}
\title{Retrieve a list of all assets}
\usage{
get_assets(con = connect_to_database())
}
\arguments{
\item{con}{The database instance. Defaults to what is returned from calling
\code{connect_to_database()}}
}
\value{
A tibble with six variables.
}
\description{
Retrieve a list of assets from PorowhitÄ Hauwha. An asset is a structure that
provides shelter for the public.
}
\examples{
con <- connect_to_database(":memory:")

assets <- porowhitahauwha:::assets
facilities_attributes <- porowhitahauwha:::facilities_attributes

DBI::dbWriteTable(con, "assets", assets)
DBI::dbWriteTable(con, "facilities_attributes", facilities_attributes)

DBI::dbDisconnect(con, shutdown=TRUE)

assets <- get_assets(":memory:")

}
