% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_assets}
\alias{get_assets}
\title{Retrieve a list of all assets}
\usage{
get_assets(..., test_db = FALSE)
}
\arguments{
\item{...}{Optional expressions to filter the asset list, defied in terms of the
variables in the returned tibble.}

\item{test_db}{Retrieve this data from the test database? Defaults to FALSE.}
}
\value{
A tibble with six variables.
}
\description{
Retrieve a list of assets from Porowhita HauwhÄ. An asset is a structure that
provides shelter for the public.
}
\examples{
# Retrieve a list of assets that are classed as rural halls from the test database
get_assets(designation == "Rural Hall", test_db = TRUE)
}
