% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_space}
\alias{insert_space}
\title{Add a New Entry to the \code{spaces} Table}
\usage{
insert_space(
  name,
  asset_id,
  designation,
  delivery_model,
  facility_ownership,
  closed = FALSE,
  leased = FALSE,
  bookable = NA,
  booking_method = NA,
  staffed = NA,
  test_db = FALSE
)
}
\arguments{
\item{name}{The name of the asset-type facility.}

\item{test_db}{Is this change being applied to a facility in the test
database (\code{TRUE}) or not (\code{FALSE})? Defaults to \code{FALSE}.}

\item{local_board}{The Local Board where the asset-type facility is located.}

\item{postal_address}{The postal address of the asset-type facility.}
}
\value{
A tibble with 1 row and 9 columns showing the newly-added asset-type
facility.
}
\description{
Add a New Entry to the \code{spaces} Table
}
