% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_contact}
\alias{update_contact}
\title{Update Partner Contact Details}
\usage{
update_contact(
  contact_name = NULL,
  email_address = NULL,
  phone_number = NULL,
  partner_id,
  test_db = FALSE
)
}
\arguments{
\item{contact_name}{The name of the partner's contact.}

\item{email_address}{The email address of the partner's contact.}

\item{phone_number}{The phone number of the partner's contact.}

\item{partner_id}{The partner id.}

\item{test_db}{Is this a connection to the test database? Defaults to
\code{FALSE}.}
}
\value{
A tibble with one row showing the updated contact details.
}
\description{
Update Partner Contact Details
}
